web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
django-import-export = "^2.6.1"
psycopg2-binary = "^2.9.1"
python-telegram-bot = "^13.7"
dramatiq = "^1.12.0"
redis = "^3.5.3"
sentry-sdk = "^1.4.3"
Pillow = "^8.4.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
